fix save booking validation, fix tests

+++ pkg/booking/booking.go
@@ -22,13 +22,13 @@ type BookingRepository interface {
 	GetTotalCount(ctx context.Context) (int64, error)
 	GetAllClassesByMemberId(ctx context.Context, memberId string) ([]class.Class, error)
 	GetAllMembersByClassId(ctx context.Context, classId string) ([]member.Member, error)
-	Save(ctx context.Context,class Booking) (int, error)
+	Save(ctx context.Context,class Booking) (int64, error)
 	Update(ctx context.Context, id string, class Booking) error
 	Delete(ctx context.Context, id string) error
 }
 
 type BookingService interface {
 	GetByDateRange(ctx context.Context,startDate, endDate string) ([]Booking, error)
-	Save(ctx context.Context,class Booking) (int, error)
+	Save(ctx context.Context,class Booking) (int64, error)
 	Update(ctx context.Context,id string, class Booking) error
 }
